+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## isoWater 1.1.2
 * Bug fixes
+* mixSource function now allows optional prior on O-isotope evaporation effect
 
 ## isoWater 1.1.1
 * iso function now checks for and removes data with missing values